faz ajustes e cria testes unitários
