+++ sme_terceirizadas/produto/api/serializers/serializers.py
@@ -342,42 +342,6 @@ class CadastroProdutosEditalSerializer(serializers.ModelSerializer):
         fields = ('uuid', 'nome', 'status')
 
 
-class CadastroProdutosEditalCreateSerializer(serializers.Serializer):
-    nome = serializers.CharField(required=True, write_only=True)
-    ativo = serializers.CharField(required=True)
-
-    def create(self, validated_data):
-        nome = validated_data['nome']
-        status = validated_data.pop('ativo')
-        ativo = False if status == 'Inativo' else True
-
-        if nome.upper() in (produto.nome.upper() for produto in NomeDeProdutoEdital.objects.all()):
-            raise serializers.ValidationError('Item já cadastrado.')
-        try:
-            produto = NomeDeProdutoEdital(nome=nome, ativo=ativo)
-            produto.save()
-            return produto
-        except Exception:
-            raise serializers.ValidationError('Erro ao criar Produto Proviniente do Edital.')
-
-    def update(self, instance, validated_data): # noqa C901
-        nome = validated_data['nome']
-        status = validated_data.pop('ativo')
-        ativo = False if status == 'Inativo' else True
-
-        if (nome.upper(), ativo) in ((produto.nome.upper(), produto.ativo)
-                                     for produto in NomeDeProdutoEdital.objects.all()):
-            raise serializers.ValidationError('Item já cadastrado.')
-
-        try:
-            instance.nome = nome.upper()
-            instance.ativo = ativo
-            instance.save()
-        except Exception as e:
-            raise serializers.ValidationError(f'Erro ao editar Produto Proviniente do Edital. {str(e)}')
-        return instance
-
-
 class ProdutosSubstitutosSerializer(serializers.ModelSerializer):
     tipo = serializers.SerializerMethodField()
